Worked on forms
Doesn't work yet. Not sure if the html template is the problem. I'm kind
of learning HTML by doing right now. I'll work on this some more
tomorrow, but we definitely need a good template for the
everything_tracker view to reference with a render call. Oh yeah, I
renamed some stuff.