Working!
Ok, I got the tracker app page to work in a very simple sense. All
patient objects in existence have their data displayed in html
paragraphs as of now, right before the add patient form. And submitting
the form just reloads the same page, but with updated patient rows.
There's still a lot to do, but this is a starting point. There was a
problem in the views file, where the line input_form.save() was causing
a duplicate patient object to be made (took me forever to figure out
what was causing that bug). I dont' know why that was doing that, but
I've commented it out, and nothing seemed to break. I'm not sure if we
really need to save the form upon input.
Also, error handling needs to be added for some notification or
prevention of users entering the date of birth field the wrong way.
Right now a DateField only accepts month/day/year, not month-day-year.
I'll try to figure out how to change that.
A worry is that if you manually refresh the page after submitting, it
will keep re-submitting that data and making more and more patient
objects. Need some way to stop that.