Error Handling on PatientForm DOB field
Ok, I made something that works. If Asher or someone wants to style the
error message, or revert it back to Django forms' built-in lists (if you
can make that work without breaking the table row), feel free. But this
is what I could do using no CSS. Rewrote the form part of the Django
template to iterate through what it would have anyway, except leaving
out any errors. I put those above. For any help text or any other
possibilites with any BoundFields that Django normally takes care of,
we'll have to manually go in and add that to the template.