Now using PatientSteps.
Changed stuff around to make PatientSteps linked to Patients. Had to
read through a lot of ForeignKey API, and mess around in the view to
make a list of tuples for each patient, with the patient object, its
last step, and its current step. I changed the template code to match.
It looks a bit messy, but it works.