Small change
Realized another exception could be thrown by a .get() call on a
QuerySet. I'm not quite sure what to do when such an error exists
though, because it would indicate a mistake in the logic of our own
code. Maybe I should just let it go and crash everything so that
we/someone can fix it..