Ordering Patients
Small change to how the patient bars are displayed. I think going by
their Time In would be a good idea, but I don't know if the nurses would
want the latest patients at the top of the screen, or the oldest
patients at the top. For now I've put the oldest patients at the top.