Changed StepOutcome relations
Made StepOutcome model have a patient foreignkey, and a one to one
relationship with a StepOutcome, which has a step foreignkey. This made
more sense to me when doing the admin page, made it possible to inline
the stepoutcomes to individual patient displays, and I think the logic
should consistently hold for database relations and working with the
view as well.