Some Work
Made PatientStep last and current boolean fields have some 'validation'
of sorts in the model's own custom save method. Also added a line in the
template to be displayed if there are no Patients. I'm going to leave
styling everything to make it look slick to Asher. Next I'm going to
work on filling in the accordion body with the appropriate forms.