Something that doesn't work.
Lots of big changes. I can elaborate on anything if need be. Right now
this seems to work up until submission. After the past two days I
consider that a success. I wanted to finish everything first, but I
figure I should check in my changes, as right now I need to go work on a
take-home test, and I may not be able to finish the accordion body.