Something that works.
Yay. It works. I did name the view appropriately however, as it is
huuuuge now. In the future we'll definitely want to clean it up, and try
to separate it into multiple views. I also want to add a static 'delete
patient' button (however I'm not sure if we'll trust the nurses to not
accidentally/maliciously delete someone else's patient), rather than
handling the submit button for the formsets to delete the patient, as I
do now. I also don't know if we really want to delete the patient from
the database after saving the patient history and stuff in some
permanent records, or if we just want to somehow make it so that the
patient doesn't show up as a bar in the tracker app anymore. Oh, I
commented out the flag column on the bars, as it wasn't working just
yet.