Switched to a Class Based View
In order to split up the GET and POST methods, as Carlos suggested.
There was a bit of funkiness in getting invalid forms to redisplay
properly, but I seem to have gotten it working using a get_context
method with optional arguments, and then checking those arguments.
Again, maybe not the cleanest way, but it works. If I have time before
tomorrow's presentation I might make the flags work, and offer a color
and a label to display on the patient bars.

I need to hear back from someone as to how we should handle removing
patients off the list. Where is our master history database that we're
storing these records to?