Changed it back.
That was ugly and hacky. Make a models.py file change so that managers
can define two or more fields as radio-buttons, then change the views.py
file to recognize that, if you want to make that change.

Also, I fixed patient outcome input so that you can override old data,
but still ensure only one value for a single stepoutcome for a single
patient.